changed DQN implementation to SB3
Removed the initial setup for a tensorflow DQN implementation, and now just initialize a standard DQN network from stable_baselines3. Thus, the NGU reward system can best be created as a wrapper for the environment, rather than an extension of the algo. There is some good explanation of how to do this on the Farama website, which is linked in the NGU file.