embedding/KNN works. BUT(read description:
There are still some problems: First of all, i'm not sure yet how to encode the observations into the correct data format to pass onto the embedding network. For now it's just the image, but ideally the direction object should also be included. I changed the embedding network kernel/stride sizes because our env observation space is way smaller than that in the paper. I debugged it for now and it seems to be working, but we are also still missing the life-long module. Enough work to do :)